move widgetGenerator files in separate folder, cleanup
- remove test route in index.js
- move widgetGenerator stuff into folder ./widget_generator
