+++ index.js
@@ -6,7 +6,7 @@
 
 var config  = require('./config.js');
 var mongo   = require('./dbConnector.js');
-var widgets = require('./widgetGenerator.js');
+var widgets = require('./widget_generator/widgetGenerator.js');
 var express = require('express');
 
 var app          = express();
@@ -33,20 +33,3 @@ mongo.connect(
 /* serve everything in the folder './public/' */
 app.use(express.static(__dirname + '/public'));
 app.use('/data', express.static(__dirname + '/data'));
-
-/* test route to check on widget creation */
-app.get('/makeWidget', function(req, res) {
-	var inPaths  = [
-		//__dirname + '/Riesel_kvr.tif',     // unprojected RGB GeoTIFF
-		//__dirname + '/o41078a1.tif',       // projected RGB GeoTIFF
-		__dirname + '/cea.tif',            // grayscale GeoTIFF
-		__dirname + '/example_01.geojson', // GeoJSON
-		__dirname + '/fig-1E.Rdata'        // SP SpatialPolygonsDataFrame
-	];
-	var outPath  = __dirname + '/data/widgets/test.html';
-
-	widgets.map(inPaths, outPath, function(err) {
-		if (err) res.send('boooh! :^(<br>' + err);
-		else     res.redirect('/data/widgets/test.html');
-	});
-});
