+++ widget_generator/makeMapWidget.r
